Ran a preliminary performance test
Looks good -- the database fetch takes 30seconds for 500,000 rows, and
including half the diagnoses and procedures takes 25seconds. Adding
all the diagnoses/procedures is still 25s (no overhead), presumably
due to the fixed short-circuiting. As a result, the entire dataset
should take 8 mins to process. There are plenty of opportunities for
optimisation.