Started refactoring the code cache
I am going to make the string lookup available everywhere, in the parser
too. The cache really needs to know about the string lookup immediately
when the cache entry is created, so that means the lookup pointer needs
to be inside the Category. I have moved the ClinicalCodeData structure and
made it the basis of the new CodeCacheEntry, which will store IDs instead of
strings.